+++ src/dosfunc.h
@@ -6,15 +6,15 @@
 #include <stddef.h>
 
 uint16 dos_alloc(const size_t paragraphs);
-int dos_free(const uint16 segment);
+int16 dos_free(const uint16 segment);
 uint16 dos_resize(const uint16 segment, uint16 newsize);
 size_t dos_getfree(void);
-int dos_loadOverlay(const char *file, const uint16 segment);
-int dos_runProgram(const char *file, const char FAR *cmdline);
-int dos_loadProgram(const char *file, const char FAR *cmdline, uint16 *cs, uint16 *ss);
-int dos_getReturnCode(void);
+int16 dos_loadOverlay(const char *file, const uint16 segment);
+int16 dos_runProgram(const char *file, const char FAR *cmdline);
+int16 dos_loadProgram(const char *file, const char FAR *cmdline, uint16 *cs, uint16 *ss);
+int16 dos_getReturnCode(void);
 uint16 dos_getProcessId(void);
-int dos_setProcessId(const uint16 pid);
+int16 dos_setProcessId(const uint16 pid);
 void dos_mcbInfo(void);
 uint16 dos_lastFreeBlock(void);
 size_t dos_envSize(void);
