+++ src/eg3dcam.h
@@ -1,8 +1,9 @@
 #ifndef F15_SE2_EG3DCAM
 #define F15_SE2_EG3DCAM
+#include "inttype.h"
 /* public interface of eg3dcam.c */
 
-void setViewRotation(int, int, int);
-void setViewPosition(int viewX, int viewY, int viewZ);
+void setViewRotation(int16, int16, int16);
+void setViewPosition(int16 viewX, int16  viewY, int16 viewZ);
 
 #endif /* F15_SE2_EG3DCAM */
