+++ src/eg3dgrid.h
@@ -1,7 +1,8 @@
 #ifndef F15_SE2_EG3DGRID
 #define F15_SE2_EG3DGRID
+#include "inttype.h"
 /* public interface of eg3dgrid.c */
 
-int process3dg(int lod, int col, int row);
+int16 process3dg(int16 lod, int16 col, int16 row);
 
 #endif /* F15_SE2_EG3DGRID */
