+++ src/eg3dproj.h
@@ -3,7 +3,7 @@
 /* public interface of eg3dproj.c */
 #include "inttype.h"
 
-uint32 scaleCoordToLod(int level, uint32 coord);
-void projectObjects(int heading, int rangeGate, long worldX, long worldY, long worldZ);
+uint32 scaleCoordToLod(int16 level, uint32 coord);
+void projectObjects(int16 heading, int16 rangeGate, long worldX, long worldY, long worldZ);
 
 #endif /* F15_SE2_EG3DPROJ */
