+++ src/egpic.h
@@ -2,6 +2,6 @@
 #define F15_SE2_EGPIC
 /* public interface of egpic.c */
 
-void openBlitClosePic(const char *filename, int page);
+void openBlitClosePic(const char *filename, int16 page);
 
 #endif /* F15_SE2_EGPIC */
